+++ src/main/java/me/zeph/relations/model/ParentsReqParam.java
@@ -2,22 +2,22 @@ package me.zeph.relations.model;
 
 public class ParentsReqParam extends OneParentReqParam {
 
-	private float m1;
-	private float m2;
+	private double m1;
+	private double m2;
 
-	public float getM1() {
+	public double getM1() {
 		return m1;
 	}
 
-	public void setM1(float m1) {
+	public void setM1(double m1) {
 		this.m1 = m1;
 	}
 
-	public float getM2() {
+	public double getM2() {
 		return m2;
 	}
 
-	public void setM2(float m2) {
+	public void setM2(double m2) {
 		this.m2 = m2;
 	}
 }
